Trait lyon_path_iterator::SvgIterator

pub trait SvgIterator: Iterator<Item=SvgEvent> + Sized {
    fn get_state(&self) -> &PathState;

    fn flattened(self, tolerance: f32) -> FlatteningIter<SvgToPathIter<Self>> { ... }
    fn path_iter(self) -> SvgToPathIter<Self> { ... }
}

An extension to the common Iterator interface, that adds information which is useful when chaining path-specific iterators.

Required Methods

fn get_state(&self) -> &PathState

The returned structure exposes the current position, the first position in the current sub-path, and the position of the last control point.

Provided Methods

fn flattened(self, tolerance: f32) -> FlatteningIter<SvgToPathIter<Self>>

Returns an iterator of FlattenedEvents, turning curves into sequences of line segments.

fn path_iter(self) -> SvgToPathIter<Self>

Returns an iterator of path events.